+++ lib/modules/manager/npm/schema.ts
@@ -1,5 +1,23 @@
 import { z } from 'zod/v4';
-import { Json, LooseRecord, Yaml } from '../../../util/schema-utils/index.ts';
+import {
+  Json,
+  LooseRecord,
+  Nullish,
+  Yaml,
+} from '../../../util/schema-utils/index.ts';
+
+// pnpm ignores registry URLs containing `${...}` env-var interpolation since v11.5.3
+function hasEnvVar(value: string): boolean {
+  return value.includes('${');
+}
+
+function withoutEnvVarRegistries(
+  registries: Record<string, string>,
+): Record<string, string> {
+  return Object.fromEntries(
+    Object.entries(registries).filter(([, url]) => !hasEnvVar(url)),
+  );
+}
 
 export const PnpmCatalogs = z.object({
   catalog: z.optional(z.record(z.string(), z.string())),
@@ -39,8 +57,10 @@ export const PnpmWorkspaceFile = Yaml.pipe(
       minimumReleaseAge: z.number().nullish(),
       minimumReleaseAgeExclude: z.array(z.string()).optional(),
       overrides: z.record(z.string(), z.string()).optional(),
-      registry: z.string().optional(),
-      registries: z.record(z.string(), z.string()).optional(),
+      registry: Nullish(z.string()),
+      registries: Nullish(
+        z.record(z.string(), z.string()).transform(withoutEnvVarRegistries),
+      ),
     })
     .and(PnpmCatalogs),
 );
